% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{circumcenter}
\alias{circumcenter}
\title{Circumcenter calculator}
\usage{
circumcenter(A, B, C)
}
\arguments{
\item{A, B, C}{Three points of the triangle}
}
\value{
Coordinates of the circumcenter in (x,y) format
}
\description{
Calculates the circumcenter of the given triangle
}
\examples{
circumcenter(c(0,0), c(1, 5), c(-3,1))
}
