% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{distance}
\alias{distance}
\title{Distance calculator}
\usage{
distance(A, B)
}
\arguments{
\item{A, B}{Two points in (x,y) format}
}
\value{
A single value, the distance between the two points
}
\description{
Calculate the distance between 2 points
}
\examples{
distance(c(0,0), c(3,4))
}
