% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{get_matrix_input}
\alias{get_matrix_input}
\title{Input of 5 points}
\usage{
get_matrix_input()
}
\value{
A 2x5 matrix with the coordinates of the 5 points
}
\description{
\code{get_matrix_input} takes 5 points as input and returns a matrix(2 row 5 col)
}
\keyword{input}
\keyword{matrix}
\keyword{points}
