% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{is_point_in_circle}
\alias{is_point_in_circle}
\title{Point position checker}
\usage{
is_point_in_circle(A, B, C, D)
}
\arguments{
\item{A, B, C}{Points of the triangle.}

\item{D}{Point whose position is to be checked.}
}
\value{
Returns TRUE if D lies in the triangle ABC, FALSE otherwise
}
\description{
Checks if point D lies in the circumcenter of triangle ABC
}
\examples{
is_point_in_circle(c(0,1), c(-1,-1), c(1,-1), c(0,0))
is_point_in_circle(c(0,1), c(-1,-1), c(1,-1), c(5,5))
}
