% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{line}
\alias{line}
\title{Line to linear equation in two variables}
\usage{
line(point, slope)
}
\arguments{
\item{point}{Given point on the line.}

\item{slope}{Given slope of the line.}
}
\value{
Ax + By = C in the form (A,B,C)
}
\description{
Calculates the necessary coefficients of the linear equation
}
\examples{
line(c(4,5), c(1,1))
}
