% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_line}
\alias{plot_line}
\title{Plot a line between 2 points}
\usage{
plot_line(A, B)
}
\arguments{
\item{A, B}{Two points each having an (x,y) value.}
}
\description{
Plots a red line between points A and B
}
\examples{
\dontrun{plot_line(c(1,2), c(3,9))}


}
\keyword{draw}
\keyword{line}
\keyword{plot}
\keyword{plotting}
