% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_points}
\alias{plot_points}
\title{Plot 5 points}
\usage{
plot_points(A, B, C, D, E)
}
\arguments{
\item{A, B, C, D, E}{Five points each having an (x,y) value.}
}
\description{
plots the 5 given point on a 2D plane
}
\examples{
plot_points(c(0,0), c(2,9), c(3,5), c(-2,-5), c(-1,3))

}
\keyword{draw}
\keyword{plot}
\keyword{plotting}
\keyword{point}
