% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_triangle}
\alias{plot_triangle}
\title{Plot a triangle between 3 points}
\usage{
plot_triangle(A, B, C)
}
\arguments{
\item{A, B, C}{three points each having an (x,y) value.}
}
\description{
Plots red lines between points A, B and C
}
\examples{
\dontrun{plot_triangle(c(0,0), c(1,1), c(-3,5))}

}
\keyword{draw}
\keyword{line}
\keyword{plot}
\keyword{plotting}
\keyword{triangle}
